Lastedits (#14)
* init branch lastedits

* norminette

* fixed exit code for syntax errors and started on error handling

* fixed some parts of errong handling

* fixed some parts of errong handling

* changed signal handling

* changed signal handling

* fixed a write byte issue, and also added a cmd check to return proper exit code

* fixed a write byte issue, and also added a cmd check to return proper exit code

* checkpoint for fighting the funcheck leaks tomorrow

* changed the way I free the ft_strarray, as I was getting a segfault, added if statements for protection, added freeing_ramp() in expand_variables error paths, added remove_last_cmd_node in init_new_cmd when current_cmd->cmd_flags fail

* revert back t working version

* for chasing leak  purposes

* finalised the error handling for different file paths

* closed 0 and 1 FDs at the end because they were leaking

* checking and correcting tester errors

* checking and correcting tester errors: cd: COMPLETE (dont handle for cd -, exit code is also working despite what tester says
echo: echo test1test2 / echo -n test1test2
env: COMPLETE
exit: exit '       -9223372036854775808'
export: COMPLETE
pwd: COMPLETE
unset: COMPLETE

* checking and correcting tester errors

* checking and correcting tester errors

* checking and correcting tester errors

* commented mini_gnl added

--------

+++ parsing/heredoc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jille <jille@student.42.fr>                +#+  +:+       +#+        */
+/*   By: esezalor <esezalor@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/10 17:56:59 by jille             #+#    #+#             */
-/*   Updated: 2026/04/10 18:31:42 by jille            ###   ########.fr       */
+/*   Updated: 2026/04/17 10:45:47 by esezalor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ void	heredoc_loop(t_cmd **cmd, int h_fd)
 		if (!line)
 		{
 			write(2, "jeis: warning: h-doc delimited by end-of-file (wanted `",
-				64);
+				56);
 			write(2, (*cmd)->heredoc_delim, strlen((*cmd)->heredoc_delim));
 			write(2, "')\n", 3);
 			break ;
@@ -80,8 +80,9 @@ int	heredoc_to_file(t_cmd **cmd)
 	static int	h_num;
 
 	if (h_num > 0)
-		if (access((*cmd)->infile, F_OK) != -1)
-			unlink((*cmd)->infile);
+		if((*cmd)->infile)
+			if (access((*cmd)->infile, F_OK) != -1)
+				unlink((*cmd)->infile);
 	if (!(*cmd)->heredoc_delim)
 		return (print_syntax_error(NULL));
 	filename = create_heredoc_file_name(h_num++);
